requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
prompt-toolkit>=3.0.43
pyyaml>=6.0.1
jinja2>=3.1.3
python-dotenv>=1.0.0
requests>=2.31.0
